feat#3. split screen and add tasks
- split screen (Home, AddTask)
- add tasks
